requirements.txt
pyglet>=2.0
python-chess>=1.999
